First cleanup pass

+++ src/main/java/mekanism/common/integration/tesla/TeslaCableIntegration.java
@@ -24,6 +24,6 @@ public class TeslaCableIntegration implements ITeslaConsumer
 	@Method(modid = "tesla")
 	public long givePower(long power, boolean simulated) 
 	{
-		return (long)Math.round(tileEntity.acceptEnergy(side, power*general.FROM_TESLA, simulated)*general.TO_TESLA);
+		return Math.round(tileEntity.acceptEnergy(side, power*general.FROM_TESLA, simulated)*general.TO_TESLA);
 	}
 }
